requirements.txt
flask==3.0.2
flask-cors==4.0.0    # optional; helps if you later call from a browser on a different port
llama-cpp-python==0.2.90
gunicorn